Refresh notification headlines after workspace renames
